pyproject.toml
[tool.poetry]
name = "airflow-with-bigquery-guide"
version = "0.1.0"
description = ""
authors = ["Yoon Hyejin <0327jane@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
mkdocs = "^1.4.2"
mkdocs-material = "^8.5.11"
apache-airflow = "^2.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
